extract Job class
